Split sink into multiple files
This commit adjusts the structure of the datadog logs sink module into multiple
files. I have some fairly large changes I'd like to experiment with here and
it's convenient to have things split up. Each file is more or less
single-concept. I have maintained the external API of this module.
