+++ src/sinks/datadog/logs/mod.rs
@@ -1,0 +1,15 @@
+use super::ApiKey;
+
+#[cfg(test)]
+mod tests;
+
+mod config;
+mod healthcheck;
+mod service;
+
+use crate::config::SinkDescription;
+use crate::sinks::datadog::logs::config::DatadogLogsConfig;
+
+inventory::submit! {
+    SinkDescription::new::<DatadogLogsConfig>("datadog_logs")
+}
